purpleleaf/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.5.0.4
bjoern==2.2.3
boto3==1.7.52
botocore==1.10.75
cairocffi==0.8.0
CairoSVG==1.0.22
celery==4.2.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
contextlib2==0.5.5
coverage==4.5.1
cryptography==2.3
cssselect2==0.2.1
Django==1.11.7
django-celery-beat==1.1.1
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-markdownx==2.0.21
django-markup==1.2
django-nose==1.4.5
django-phonenumber-field==2.0.0
django-redis==4.10.0
djangorestframework==3.7.7
docutils==0.14
enum34==1.1.6
futures==3.2.0
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
kombu==4.2.1
Markdown==2.6.11
nose==1.3.7
olefile==0.44
pdfrw==0.4
phonenumberslite==8.9.10
Pillow==4.3.0
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyotp==2.2.6
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.7.3
pytz==2017.3
raven==6.9.0
redis==2.10.6
reportlab==3.4.0
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
tinycss2==0.6.1
twilio==6.14.4
urllib3==1.23
vine==1.1.4
WeasyPrint==0.42
webencodings==0.5.1
whois==0.7
xhtml2pdf==0.2b1
